Style BMI font color, extract calculate function from PatientCard to Calculators.js
